requirements.txt
Flask==1.1.2
gunicorn==20.1.0
pandas>=0.19
numpy>=1.9.2
scikit-learn>=0.18
scikit-surprise==1.1.1
matplotlib>=1.4.3
seaborn==0.11.1